moved files related functions into file_utils
